Moved Footer component to its own folder
